+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/index/ContextIndex.java
@@ -16,6 +16,7 @@
  */
 package com.wazuh.contentmanager.index;
 
+import com.wazuh.contentmanager.client.CTIClient;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.opensearch.action.index.IndexRequest;
@@ -28,10 +29,16 @@ import org.opensearch.core.action.ActionListener;
 import org.opensearch.core.xcontent.ToXContent;
 import org.opensearch.index.query.QueryBuilders;
 import org.opensearch.index.query.TermQueryBuilder;
+import org.opensearch.search.SearchHit;
 import org.opensearch.search.builder.SearchSourceBuilder;
 
 import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 import com.wazuh.contentmanager.model.ctiapi.ConsumerInfo;
 
@@ -41,6 +48,7 @@ public class ContextIndex {
 
     /** The name of the Contexts index */
     public static final String CONTEXTS_INDEX = "wazuh-context";
+    public static final Long TIMEOUT = 10L;
 
     private final Client client;
 
@@ -70,7 +78,7 @@ public class ContextIndex {
                             .id(consumerInfo.getContext())
                             .create(true);
         } catch (IOException e) {
-            log.error("Error creating Catalog IndexRequest: {}", e.getMessage());
+            log.error("Failed to create JSON content builder: {}", e.getMessage());
         }
 
         this.client.index(
@@ -79,15 +87,15 @@ public class ContextIndex {
                     @Override
                     public void onResponse(IndexResponse indexResponse) {
                         log.info(
-                                "CTI Catalog Context {}, indexed with status {}",
-                                indexResponse.getId(),
-                                indexResponse.status());
+                                "Consumer [{}], indexed with status {}",
+                                consumerInfo.getContext(),
+                                indexResponse.getId());
                     }
 
                     @Override
                     public void onFailure(Exception e) {
                         log.error(
-                                "Failed to index CTI Catalog Context {}, Exception: {}",
+                                "Failed to index Consumer [{}] information due to: {}",
                                 consumerInfo.getContext(),
                                 e);
                     }
@@ -131,5 +139,18 @@ public class ContextIndex {
                     }
                 });
         return future;
+    }
+
+    @SuppressWarnings("unchecked")
+    ConsumerInfo getConsumer(String context, String consumer) {
+        try {
+            SearchResponse searchResponse = get(context).get(TIMEOUT, TimeUnit.SECONDS);
+            Map<String, Object> source = (Map<String, Object>) searchResponse.getHits().getHits()[0].getSourceAsMap().get(consumer);
+            Long last_offset = (Long) source.get(ConsumerInfo.LAST_OFFSET);
+            String snapshot = (String) source.get(ConsumerInfo.LAST_SNAPSHOT_LINK);
+            return new ConsumerInfo(consumer, context, last_offset, snapshot);
+        } catch (InterruptedException | ExecutionException | TimeoutException e) {
+            log.error("Failed to retrieve context [{}], consumer [{}]: {}", context, consumer, e.getMessage());
+        }
     }
 }